requirements.txt
Django==1.7.1
django-model-utils==2.2
wsgiref==0.1.2
